feat: adding loading to cards
