+++ src/api/dashboardStats.js
@@ -2,7 +2,6 @@ import axios from "axios";
 
 const BASE_URL = process.env.REACT_APP_API_URL;
 
-
 const countResponsesByStudentLength = (responses) => {
   const countMap = {};
